requirements.txt
# Core
yt-dlp>=2023.3.4
openai-whisper>=20230314
deep-translator>=1.11.4
ffmpeg-python>=0.2.0
pysrt>=1.1.2

# AI Subtitle Optimization
zhipuai>=2.0.0        # GLM API (Primary)
anthropic>=0.18.0     # Claude API (Alternative)
google-generativeai>=0.3.0  # Gemini API (Alternative)

# Audio / Dubbing
edge-tts>=6.1.9       # Microsoft Edge TTS
pydub>=0.25.1         # Audio processing

# Utilities
python-dotenv>=1.0.0  # Environment variable management
rich>=13.0.0          # Terminal progress bars and styled output
